[REFACTOR] Food const 폴더 위치 변경
